feat: move core logic to context
